+++ src/pages/LoginPage.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '@/components/LoginForm';
+import LoginForm from '@/components/Login/LoginForm';
 
 export default function LoginPage() {
   return (
